+++ core/src/lib.rs
@@ -18,6 +18,7 @@
 //!   failure.
 
 pub mod casp;
+pub mod creators;
 pub mod curse;
 pub mod db;
 pub mod dbpf;